# go-api/internal/http/openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "Fullstack Oracle API", "version": "1.0.0" },
  "servers": [{ "url": "/" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" },
          "fields": { "type": "object", "additionalProperties": { "type": "string" } },
          "request_id": { "type": "string" }
        }
      },
      "Tokens": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "refresh_token": { "type": "string" }
        },
        "required": ["access_token", "refresh_token"]
      },
      "Me": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "role": { "type": "string" }
        }
      },
      "Item": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "price": { "type": "number", "format": "float" },
          "created_at": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "name", "price", "created_at"]
      },
      "CreateItemDTO": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "price": { "type": "number", "minimum": 0 }
        },
        "required": ["name", "price"]
      },
      "PagedItems": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/Item" } },
          "page": { "type": "integer" },
          "size": { "type": "integer" },
          "total": { "type": "integer", "format": "int64" }
        },
        "required": ["items", "page", "size", "total"]
      }
    }
  },
  "paths": {
    "/health": {
      "get": { "tags": ["health"], "responses": { "200": { "description": "OK" } } }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "email": { "type": "string", "format": "email" }, "password": { "type": "string" } },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Tokens", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tokens" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "parameters": [{ "in": "header", "name": "X-Refresh-Token", "required": false, "schema": { "type": "string" } }],
        "requestBody": {
          "required": false,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "refresh_token": { "type": "string" } } } } }
        },
        "responses": {
          "200": { "description": "Tokens", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tokens" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Me" } } } }, "401": { "description": "Unauthorized" } }
      }
    },
    "/items/": {
      "get": {
        "tags": ["items"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "in": "query", "name": "size", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } },
          { "in": "query", "name": "sort", "schema": { "type": "string" }, "description": "field,dir e.g. price,asc" },
          { "in": "query", "name": "q", "schema": { "type": "string" }, "description": "search by name (ILIKE)" }
        ],
        "responses": {
          "200": { "description": "Paged items", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PagedItems" } } } },
          "401": { "description": "Unauthorized" },
          "429": { "description": "Rate limited", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "post": {
        "tags": ["items"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateItemDTO" } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/items/{id}": {
      "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer", "format": "int64" } }],
      "get": {
        "tags": ["items"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } }, "headers": { "ETag": { "schema": { "type": "string" } } } },
          "304": { "description": "Not Modified" },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "put": {
        "tags": ["items"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateItemDTO" } } } },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["items"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "No Content" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    }
  }
}